+++ CHANGELOG.md
@@ -7,6 +7,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 
+## [0.6.0] - 2016-11-17
+### Changed
+- DB base class changes to support field configuration across all database types
+- Change base create and update table methods to allow for native settings
+
+### Fixed
+- DF-853 Fixed broken SNS APIs
+- Trim out parenthesis of single filter
+- Type marshalling for boolean
+
 ## [0.5.0] - 2016-10-03
 ### Added
 - Beta release of Redshift service support
@@ -14,8 +24,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - DF-641 Download files in chunks
 - DF-826 Protecting secret key using service config rework from df-core
-
-### Fixed
 
 ## [0.4.0] - 2016-08-21
 ### Changed
@@ -55,7 +63,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2015-10-24
 First official release working with the new [df-core](https://github.com/dreamfactorysoftware/df-core) library.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-aws/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-aws/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/dreamfactorysoftware/df-aws/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/dreamfactorysoftware/df-aws/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/dreamfactorysoftware/df-aws/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/dreamfactorysoftware/df-aws/compare/0.3.0...0.3.1